# internal/delivery/http/v1/handler_stubs.go
package v1

import (
	"net/http"
	"rare-merch-exchange/internal/domain/auction"
	"rare-merch-exchange/internal/domain/contest"
	"rare-merch-exchange/internal/domain/item"
	"rare-merch-exchange/internal/domain/lottery"
	"rare-merch-exchange/internal/domain/transaction"
	"rare-merch-exchange/internal/domain/verification"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"
)

// ItemHandler handler для товаров
type ItemHandler struct {
	useCase item.ItemUseCase
	logger  *zap.Logger
}

func NewItemHandler(useCase item.ItemUseCase, logger *zap.Logger) *ItemHandler {
	return &ItemHandler{useCase: useCase, logger: logger}
}

// CreateItem создание товара
// @Summary Создать товар
// @Tags items
// @Accept json
// @Produce json
// @Param request body item.CreateItemRequest true "Данные товара"
// @Success 201 {object} item.Item
// @Failure 400 {object} ErrorResponse
// @Failure 401 {object} ErrorResponse
// @Security BearerAuth
// @Router /items [post]
func (h *ItemHandler) CreateItem(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

// GetItem получение товара по ID
// @Summary Получить товар по ID
// @Tags items
// @Accept json
// @Produce json
// @Param id path string true "ID товара"
// @Success 200 {object} item.Item
// @Failure 404 {object} ErrorResponse
// @Router /items/{id} [get]
func (h *ItemHandler) GetItem(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ItemHandler) UpdateItem(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ItemHandler) DeleteItem(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

// SearchItems поиск товаров
// @Summary Поиск товаров
// @Tags items
// @Accept json
// @Produce json
// @Param query query string false "Поисковый запрос"
// @Param category query string false "Категория"
// @Param limit query int false "Лимит результатов" default(20)
// @Param offset query int false "Смещение" default(0)
// @Success 200 {array} item.Item
// @Failure 400 {object} ErrorResponse
// @Router /items [get]
func (h *ItemHandler) SearchItems(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ItemHandler) GetFeaturedItems(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ItemHandler) GetItemsByCategory(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ItemHandler) SubmitForVerification(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ItemHandler) GetAuthenticityStatus(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

// AuctionHandler handler для аукционов
type AuctionHandler struct {
	useCase auction.AuctionUseCase
	logger  *zap.Logger
}

func NewAuctionHandler(useCase auction.AuctionUseCase, logger *zap.Logger) *AuctionHandler {
	return &AuctionHandler{useCase: useCase, logger: logger}
}

// CreateAuction создание аукциона
// @Summary Создать аукцион
// @Tags auctions
// @Accept json
// @Produce json
// @Param request body auction.CreateAuctionRequest true "Данные аукциона"
// @Success 201 {object} auction.Auction
// @Failure 400 {object} ErrorResponse
// @Failure 401 {object} ErrorResponse
// @Security BearerAuth
// @Router /auctions [post]
func (h *AuctionHandler) CreateAuction(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

// GetAuction получение аукциона по ID
// @Summary Получить аукцион по ID
// @Tags auctions
// @Accept json
// @Produce json
// @Param id path string true "ID аукциона"
// @Success 200 {object} auction.Auction
// @Failure 404 {object} ErrorResponse
// @Router /auctions/{id} [get]
func (h *AuctionHandler) GetAuction(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) UpdateAuction(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) CancelAuction(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) SearchAuctions(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) GetActiveAuctions(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) GetEndingSoonAuctions(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

// PlaceBid размещение ставки
// @Summary Разместить ставку на аукцион
// @Tags auctions
// @Accept json
// @Produce json
// @Param id path string true "ID аукциона"
// @Param request body auction.PlaceBidRequest true "Данные ставки"
// @Success 201 {object} auction.Bid
// @Failure 400 {object} ErrorResponse
// @Failure 401 {object} ErrorResponse
// @Security BearerAuth
// @Router /auctions/{id}/bid [post]
func (h *AuctionHandler) PlaceBid(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) GetBids(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) CreateAutoBid(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) UpdateAutoBid(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) DeleteAutoBid(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *AuctionHandler) GetAutoBids(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

// LotteryHandler handler для лотерей
type LotteryHandler struct {
	useCase lottery.LotteryUseCase
	logger  *zap.Logger
}

func NewLotteryHandler(useCase lottery.LotteryUseCase, logger *zap.Logger) *LotteryHandler {
	return &LotteryHandler{useCase: useCase, logger: logger}
}

func (h *LotteryHandler) CreateLottery(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *LotteryHandler) GetLottery(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *LotteryHandler) UpdateLottery(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *LotteryHandler) CancelLottery(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *LotteryHandler) SearchLotteries(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *LotteryHandler) GetActiveLotteries(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *LotteryHandler) GetEndingSoonLotteries(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *LotteryHandler) BuyTickets(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *LotteryHandler) GetUserTickets(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *LotteryHandler) VerifyDraw(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

// ContestHandler handler для конкурсов
type ContestHandler struct {
	useCase contest.ContestUseCase
	logger  *zap.Logger
}

func NewContestHandler(useCase contest.ContestUseCase, logger *zap.Logger) *ContestHandler {
	return &ContestHandler{useCase: useCase, logger: logger}
}

func (h *ContestHandler) CreateContest(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) GetContest(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) UpdateContest(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) CancelContest(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) SearchContests(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) GetActiveContests(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) GetEndingSoonContests(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) SubmitContest(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) GetSubmissions(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) VoteContest(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *ContestHandler) GetVotes(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

// TransactionHandler handler для транзакций
type TransactionHandler struct {
	useCase transaction.TransactionUseCase
	logger  *zap.Logger
}

func NewTransactionHandler(useCase transaction.TransactionUseCase, logger *zap.Logger) *TransactionHandler {
	return &TransactionHandler{useCase: useCase, logger: logger}
}

func (h *TransactionHandler) CreateTransaction(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *TransactionHandler) GetTransaction(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *TransactionHandler) GetUserTransactions(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *TransactionHandler) GetUserTransactionsAsBuyer(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *TransactionHandler) GetUserTransactionsAsSeller(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *TransactionHandler) UpdateShipping(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *TransactionHandler) ConfirmDelivery(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *TransactionHandler) CreateDispute(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *TransactionHandler) GetUserDisputes(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *TransactionHandler) GetDispute(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

// VerificationHandler handler для верификации
type VerificationHandler struct {
	useCase verification.VerificationUseCase
	logger  *zap.Logger
}

func NewVerificationHandler(useCase verification.VerificationUseCase, logger *zap.Logger) *VerificationHandler {
	return &VerificationHandler{useCase: useCase, logger: logger}
}

func (h *VerificationHandler) SubmitVerification(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *VerificationHandler) GetVerification(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *VerificationHandler) GetPendingVerifications(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *VerificationHandler) GetVerificationStats(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *VerificationHandler) VerifyItem(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *VerificationHandler) GetExperts(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}

func (h *VerificationHandler) RegisterExpert(c *gin.Context) {
	// TODO: реализовать
	c.JSON(http.StatusNotImplemented, ErrorResponse{
		Error:   "not_implemented",
		Message: "Функция пока не реализована",
	})
}